requirements.txt
discord
asyncio
valve
python-valve
ipinfo
requests
python-a2s